Hangart/requirements.txt
Django>=5.2.4
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
django-cors-headers>=4.3.0
django-filter>=23.5
Pillow>=10.0.0
